% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_G_alumni_109.R
\docType{data}
\name{a_G_alumni_109}
\alias{a_G_alumni_109}
\title{Column-normalized adjacency matrix representing a subsample of the alumni network of the 109th U.S. Congress.}
\format{an object of class \code{Matrix} with 429 rows and 429 columns}
\usage{
data("a_G_alumni_109")
}
\description{
Column-normalized adjacency matrix representing a subsample of the alumni network of the 109th U.S. Congress.
}
\references{
Battaglini M., E. Patacchini (2018), "Influencing Connected Legislators", Journal of Political Economy, forthcoming.\cr
}
\keyword{datasets}
