% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_G_alumni_111.R
\docType{data}
\name{a_G_alumni_111}
\alias{a_G_alumni_111}
\title{Column-normalized adjacency matrix representing the alumni network of the 111th U.S. Congress.}
\format{
an object of class \code{Matrix} with 426 rows and 426 columns
}
\usage{
data("a_G_alumni_111")
}
\description{
Column-normalized adjacency matrix representing the alumni network of the 111th U.S. Congress.
}
\references{
Battaglini M., E. Patacchini (2018), "Influencing Connected Legislators", Journal of Political Economy, forthcoming.\cr
}
\keyword{datasets}
