% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_G_alumni_112.R
\docType{data}
\name{a_G_alumni_112}
\alias{a_G_alumni_112}
\title{Column-normalized adjacency matrix representing the alumni network of the 112th U.S. Congress.}
\format{an object of class \code{Matrix} with 422 rows and 422 columns}
\usage{
data("a_G_alumni_112")
}
\description{
Column-normalized adjacency matrix representing the alumni network of the 112th U.S. Congress.
}
\references{
Battaglini M., E. Patacchini (2018), "Influencing Connected Legislators", Journal of Political Economy, forthcoming.\cr
}
\keyword{datasets}
