% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_G_alumni_113.R
\docType{data}
\name{a_G_alumni_113}
\alias{a_G_alumni_113}
\title{Column-normalized adjacency matrix representing the alumni network of the 113th U.S. Congress.}
\format{an object of class \code{Matrix} with 407 rows and 407 columns}
\usage{
data("a_G_alumni_113")
}
\description{
Column-normalized adjacency matrix representing the alumni network of the 113th U.S. Congress.
}
\references{
Battaglini M., E. Patacchini (2018), "Influencing Connected Legislators", Journal of Political Economy, forthcoming.
}
\keyword{datasets}
