% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_G_party_111.R
\docType{data}
\name{a_G_party_111}
\alias{a_G_party_111}
\title{Column-normalized adjacency matrix representing the alumni network of the 111th U.S. Congress weighted by party affiliation. Weights take values of 1 if the linked congressmen have the same school and party, 0.5 if they have the same school and a different party, 0 if they have a different school and the same party, and -0.5 if they have a different school and a different party.}
\format{an object of class \code{Matrix} with 426 rows and 426 columns}
\usage{
data("a_G_party_111")
}
\description{
Column-normalized adjacency matrix representing the alumni network of the 111th U.S. Congress weighted by party affiliation. Weights take values of 1 if the linked congressmen have the same school and party, 0.5 if they have the same school and a different party, 0 if they have a different school and the same party, and -0.5 if they have a different school and a different party.
}
\references{
Battaglini M., E. Patacchini (2018), "Influencing Connected Legislators", Journal of Political Economy, forthcoming.\cr
}
\keyword{datasets}
