% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_alumni_110.R
\docType{data}
\name{G_alumni_110}
\alias{G_alumni_110}
\title{Column-normalized adjacency matrix representing the alumni network of the 110th U.S. Congress.}
\format{an object of class \code{Matrix} with 442 rows and 442 columns}
\usage{
data("G_alumni_110")
}
\description{
Column-normalized adjacency matrix representing the alumni network of the 110th U.S. Congress.
}
\references{
Battaglini M., V. Leone Sciabolazza, E. Patacchini (2018), "Effectiveness of Connected Legislators",  Mimeo.
}
\keyword{datasets}
