% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_alumni_112.R
\docType{data}
\name{G_alumni_112}
\alias{G_alumni_112}
\title{Column-normalized adjacency matrix representing the alumni network of the 112th U.S. Congress.}
\format{an object of class \code{Matrix} with 422 rows and 422 columns}
\usage{
data("G_alumni_112")
}
\description{
Column-normalized adjacency matrix representing the alumni network of the 112th U.S. Congress.
}
\references{
Battaglini M., V. Leone Sciabolazza, E. Patacchini (2018), "Effectiveness of Connected Legislators",  Mimeo.
}
\keyword{datasets}
