% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_cosponsor_110.R
\docType{data}
\name{G_cosponsor_110}
\alias{G_cosponsor_110}
\title{Column-normalized adjacency matrix representing the cosponsorship network of the 110th U.S. Congress.}
\format{an object of class \code{Matrix} with 448 rows and 448 columns}
\usage{
data("G_cosponsor_110")
}
\description{
Column-normalized adjacency matrix representing the cosponsorship network of the 110th U.S. Congress.
}
\references{
Battaglini M., V. Leone Sciabolazza, E. Patacchini (2018), "Effectiveness of Connected Legislators", Mimeo.\cr
}
\keyword{datasets}
