% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_cosponsor_113.R
\docType{data}
\name{G_cosponsor_113}
\alias{G_cosponsor_113}
\title{Column-normalized adjacency matrix representing the cosponsorship network of the 113th U.S. Congress.}
\format{an object of class \code{Matrix} with 416 rows and 416 columns}
\usage{
data("G_cosponsor_113")
}
\description{
Column-normalized adjacency matrix representing the cosponsorship network of the 113th U.S. Congress.
}
\references{
Battaglini M., V. Leone Sciabolazza, E. Patacchini (2018), "Effectiveness of Connected Legislators", Mimeo.\cr
}
\keyword{datasets}
