% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_model_A_test.R
\docType{data}
\name{G_model_A_test}
\alias{G_model_A_test}
\title{Column-normalized adjacency matrix representing a subsample of the alumni network of the 111th U.S. Congress.}
\format{
Data set for runtime executions. An object of class \code{Matrix} with 42 rows and 42 columns
}
\usage{
data("G_model_A_test")
}
\description{
Column-normalized adjacency matrix representing a subsample of the alumni network of the 111th U.S. Congress.
}
\references{
Battaglini M., E. Patacchini (2018), "Influencing Connected Legislators", Journal of Political Economy, forthcoming.\cr
}
\keyword{datasets}
