% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot}
\alias{boot}
\alias{boot.econet}
\title{boot: Bootstrap residuals with cross-sectional dependence}
\usage{
\method{boot}{econet}(object, hypothesis = c("lim", "het", "het_l",
  "het_r", "par"), group = NULL, niter, weights, ...)
}
\arguments{
\item{object}{an object of class \code{econet} obtained using 'NLLS' estimation.}

\item{hypothesis}{string. One of \code{c("lim","het", "het_l", "het_r", "par")}.}

\item{group}{\code{NULL} or vector of positive integers specifying the indices for resampling within groups.}

\item{niter}{number of required iterations.}

\item{weights}{logical. It is \code{TRUE} if the object \code{object} was estimated using \code{weights}, and \code{FALSE} otherwise.}

\item{...}{additional parameters}
}
\value{
a numeric vector containing bootstrapped standard errors  (see Anselin, 1990). If the procedure is not feasible, it returns a vector of NAs.
}
\description{
boot: Bootstrap residuals with cross-sectional dependence
}
\details{
For additional details, see the vignette
Warning: This function is available only when \code{net_dep} is run with \code{estimation == "NLLS"}
}
\examples{
\donttest{
# Load data
data("db_cosponsor")
data("G_alumni_111")
db_model_B <- db_cosponsor
G_model_B <- G_cosponsor_111
G_exclusion_restriction <- G_alumni_111
are_factors <- c("party", "gender", "nchair")
db_model_B[are_factors] <- lapply(db_model_B[are_factors], factor)

# Specify formula
f_model_B <- formula("les ~gender + party + nchair")

# Specify starting values
starting <- c(alpha = 0.23952,
              beta_gender1 = -0.22024,
              beta_party1 = 0.42947,
              beta_nchair1 = 3.09615,
              phi = 0.40038,
              unobservables = 0.07714)

# object Linear-in-means model
lim_model_B <- net_dep(formula = f_model_B, data = db_model_B,
                       G = G_model_B, model = "model_B", estimation = "NLLS",
                       hypothesis = "lim", endogeneity = TRUE, correction = "heckman",
                       first_step = "standard",
                       exclusion_restriction = G_exclusion_restriction,
                       start.val = starting)
# Bootstrap
# Warning: this may take a very long time to run.
# Decrease the number of iterations to reduce runtime.
boot_lim_estimate <- boot(object = lim_model_B, hypothesis = "lim",
                          group = NULL, niter = 1000, weights = FALSE)
boot_lim_estimate
}
}
\references{
Anselin, L., 1990, "Some robust approach to testing and estimation in spatial econometrics", Regional Science and Urban Economics, 20, 141-163.
}
\seealso{
\code{\link{net_dep}}
}
